Hashmap
HashMapp only takes a file with two columns and makes a hash map. I am
trying to make the object map work but it keeps giving me errors. So
for now I am going to push this and I’ll modify it soon.
ObjUser is just the user class for the user object.